Updated week chart to sort by year
